lift user input state up
- lift user input state up from UserInput to the App component and accept onChangeInput prop
- add Results component
