+++ src/components/Results.jsx
@@ -1,0 +1,9 @@
+// This component displays the investment results in a table format.
+// It will get the user input as an input here instead of the UserInput component,
+// so that the results can be calculated and displayed dynamically.
+export default function Results({ input }) {
+    console.log(input);
+    return (
+        <p>Results ... </p>        
+    );
+}
